Extract cohesive Rust policy and HTTP modules within line limits
